backend/requirements.txt
python-dotenv
fastapi
Request
google.generativeai
uvicorn